pulled deprecation into its own submodule
