+++ librosa/util/__init__.py
@@ -81,6 +81,7 @@ Deprecated
 from .utils import *  # pylint: disable=wildcard-import
 from .files import *  # pylint: disable=wildcard-import
 from .feature_extractor import *  # pylint: disable=wildcard-import
+from .deprecation import * # pylint: disable=wildcard-import
 from . import decorators
 from . import exceptions
